fix(fsoe): widen master param indices to uint16_t and limit slave paramBuf allocation
The master's param_tx_idx_ and param_rx_idx_ were uint8_t (max 255) but
PARAM_BUF_SIZE is 262 (6-byte header + 256 app params). With >249 app
parameters, the indices would silently truncate on assignment, causing
the master to never complete the parameter transfer (stuck in Parameter
state) or access incorrect buffer offsets. Widen both to uint16_t to
match the slave's paramRxIdx_/paramTxIdx_ types and PARAM_BUF_SIZE.

Also add a PARAM_BUF_SIZE limit in the slave's processParameter: if a
malicious master sends an app_param_len that would make total_len exceed
PARAM_BUF_SIZE, the slave now rejects the frame with ParameterError
instead of calling paramBuf_.resize() with a potentially huge size
(up to 65541 bytes for app_param_len=65535).

+++ include/tether/fsoe/FSoEMasterConnection.hpp
@@ -407,8 +407,8 @@ private:
     // See: https://techoverflow.net/2026/08/12/fsoe-parameter-pdu-master-and-slave-structure/
     uint8_t current_param_index_ = 0;   ///< Legacy: parameter cycle count
     uint16_t parameter_crc_ = 0;        ///< Legacy: parameter CRC (diagnostics)
-    uint8_t param_tx_idx_ = 0;          ///< TX byte offset into param_tx_buf_
-    uint8_t param_rx_idx_ = 0;          ///< RX echo byte offset into param_rx_buf_
+    uint16_t param_tx_idx_ = 0;         ///< TX byte offset into param_tx_buf_
+    uint16_t param_rx_idx_ = 0;         ///< RX echo byte offset into param_rx_buf_
     static constexpr uint16_t PARAM_BUF_SIZE = 256 + 6;  ///< 6-byte header + max app params
     std::vector<uint8_t> param_tx_buf_;  ///< Full parameter payload to send
     std::vector<uint8_t> param_rx_buf_;  ///< Accumulated echo from slave
